requirements.txt
asgiref==3.4.1
Django==4.0
djangorestframework==3.13.0
joblib==1.1.0
lightgbm==3.3.1
numpy==1.21.4
pandas==1.3.5
Pillow==8.4.0
python-dateutil==2.8.2
pytz==2021.3
scikit-learn==1.0.1
scipy==1.7.3
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.0.0
tzdata==2021.5
django-filter==2.4.0
django-heroku==0.3.1
django-storages==1.10.1
django-utils-six==2.0
whitenoise==4.0
gunicorn==20.0.4
psycopg2==2.8.6
psycopg2-binary==2.8.6
